解决所有警告，将NSInteger类型转为int类型。年月日用int足够。少占用4个字节

+++ JNWarmTips/Calendar/ViewController/ViewController+CalculateCalendar.h
@@ -7,7 +7,7 @@
 #import "ViewController.h"
 
 @interface ViewController (CalculateCalendar)
-- (NSMutableArray *) getAllDaysOfMonth:(NSInteger)month InYear:(NSInteger)year;
-- (NSString *) getLastMonth:(NSInteger)currentMonth currentYear:(NSInteger)currentYear;
-- (NSString *) getNextMonth:(NSInteger)currentMonth currentYear:(NSInteger)currentYear;
+- (NSMutableArray *) getAllDaysOfMonth:(int)month InYear:(int)year;
+- (NSString *) getLastMonth:(int)currentMonth currentYear:(int)currentYear;
+- (NSString *) getNextMonth:(int)currentMonth currentYear:(int)currentYear;
 @end
